Create BDT2_2020_Varghese Betsy_Olivera Andres_Gutierrez Sara.py
Business Context:
The unique positioning of the company in the French market was being challenged by a competitor offering the same services for lower prices.

Aim of the Project:
To predict the proption of customers who would churn. i.e. switch to the competitor's services.

Dataset Description:
The data belonged to a French B2B food catering company. It included details about client subscriptions, food deliveries, and complaints.